code refactor

+++ core/vm/interpreter.cpp
@@ -448,7 +448,7 @@ void StackInterpreter::Execute(size_t* op_stack, long* stack_pos, long i,
 #ifdef _DEBUG
       wcout << L"stack oper: I2F; call_pos=" << (*call_stack_pos) << endl;
 #endif
-      PushFloat(PopInt(op_stack, stack_pos), op_stack, stack_pos);
+      PushFloat((long)PopInt(op_stack, stack_pos), op_stack, stack_pos);
       break;
 
     case F2I:
@@ -623,7 +623,7 @@ void StackInterpreter::Execute(size_t* op_stack, long* stack_pos, long i,
 #endif
 
 #ifdef _WIN32
-      left = PopInt(op_stack, stack_pos);
+      left = (long)PopInt(op_stack, stack_pos);
       Sleep(left);
 #else
       left = PopInt(op_stack, stack_pos);
@@ -716,7 +716,7 @@ void StackInterpreter::StorClsInstIntVar(StackInstr* instr, size_t* &op_stack, l
     ::exit(1);
 #endif
   }
-  long mem = PopInt(op_stack, stack_pos);
+  size_t mem = PopInt(op_stack, stack_pos);
   cls_inst_mem[instr->GetOperand()] = mem;
 }
 
@@ -754,8 +754,8 @@ void StackInterpreter::ShlInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: SHL_INT; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left << right, op_stack, stack_pos);
 }
 
@@ -764,8 +764,8 @@ void StackInterpreter::ShrInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: SHR_INT; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left >> right, op_stack, stack_pos);
 }
 
@@ -802,8 +802,8 @@ void StackInterpreter::AndInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: AND; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left && right, op_stack, stack_pos);
 }
 
@@ -812,8 +812,8 @@ void StackInterpreter::OrInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: OR; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left || right, op_stack, stack_pos);
 }
 
@@ -822,8 +822,8 @@ void StackInterpreter::AddInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: ADD; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left + right, op_stack, stack_pos);
 }
 
@@ -842,8 +842,8 @@ void StackInterpreter::SubInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: SUB; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left - right, op_stack, stack_pos);
 }
 
@@ -862,8 +862,8 @@ void StackInterpreter::MulInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: MUL; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left * right, op_stack, stack_pos);
 }
 
@@ -872,8 +872,8 @@ void StackInterpreter::DivInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: DIV; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left / right, op_stack, stack_pos);
 }
 
@@ -902,8 +902,8 @@ void StackInterpreter::ModInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: MOD; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left % right, op_stack, stack_pos);
 }
 
@@ -912,8 +912,8 @@ void StackInterpreter::BitAndInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: BIT_AND; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left & right, op_stack, stack_pos);
 }
 
@@ -922,8 +922,8 @@ void StackInterpreter::BitOrInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: BIT_OR; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left | right, op_stack, stack_pos);
 }
 
@@ -932,8 +932,8 @@ void StackInterpreter::BitXorInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: BIT_XOR; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left ^ right, op_stack, stack_pos);
 }
 
@@ -942,8 +942,8 @@ void StackInterpreter::LesEqlInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: LES_EQL; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left <= right, op_stack, stack_pos);
 }
 
@@ -952,8 +952,8 @@ void StackInterpreter::GtrEqlInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: GTR_EQL; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left >= right, op_stack, stack_pos);
 }
 
@@ -982,8 +982,8 @@ void StackInterpreter::EqlInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: EQL; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left == right, op_stack, stack_pos);
 }
 
@@ -992,8 +992,8 @@ void StackInterpreter::NeqlInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: NEQL; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left != right, op_stack, stack_pos);
 }
 
@@ -1002,8 +1002,8 @@ void StackInterpreter::LesInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: LES; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left < right, op_stack, stack_pos);
 }
 
@@ -1012,8 +1012,8 @@ void StackInterpreter::GtrInt(size_t* &op_stack, long* &stack_pos)
 #ifdef _DEBUG
   wcout << L"stack oper: GTR; call_pos=" << (*call_stack_pos) << endl;
 #endif
-  const long left = PopInt(op_stack, stack_pos);
-  const long right = PopInt(op_stack, stack_pos);
+  const long left = (long)PopInt(op_stack, stack_pos);
+  const long right = (long)PopInt(op_stack, stack_pos);
   PushInt(left > right, op_stack, stack_pos);
 }
